requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
fasttext-wheel>=0.9.2
requests>=2.31.0
